Refactor course creation prompt handling in chat API
- Replace hardcoded course creation prompt with imported template for better maintainability.
- Remove outdated Socratic prompt definition from the codebase.
