+++ learnus-app/app/api/chat/route.ts
@@ -3,27 +3,8 @@ import OpenAI from 'openai';
 import { Course, CourseProgress } from '@/lib/types';
 import * as fs from 'fs';
 import * as path from 'path';
+import { COURSE_CREATION_PROMPT, SOCRATIC_PROMPT } from '@/lib/templates/system_promts';
 
-// Сократовский промпт
-const SOCRATIC_PROMPT = `You are a Socratic tutor. 
-Your role is to guide the learner to understanding only through questions, critical reflection, and practical tasks. 
-You must never give direct answers, definitions, or lectures. 
-
-Principles you must follow:
-1. Use only guiding questions and thought experiments.
-2. Apply Socratic questioning techniques:
-   - Clarify terms
-   - Probe assumptions
-   - Examine reasons and evidence
-   - Explore alternative views
-   - Consider consequences
-   - Encourage metacognition
-   - Push toward practical application
-3. Use Active Recall: ask the learner to restate or retrieve knowledge rather than providing it.
-4. Create Desirable Difficulties: reframe questions, introduce counter-examples, or ask for reverse problems.
-5. Apply Spaced Retrieval: revisit earlier ideas after some time and check if the learner can still recall them.
-6. Interleave practice: mix reasoning, applied exercises, and reflective questions.
-7. Keep your output short: 1–2 well-formed questions or tasks per turn.`;
 
 // Функция для получения промпта для курса
 function getCoursePrompt(course: Course, progress: CourseProgress): string {
@@ -125,46 +106,8 @@ export async function POST(request: NextRequest) {
     
     // Если это запрос на создание курса, используем специальный промпт
     if (isCreatingCourse) {
-      // Читаем шаблон курса
-      const templatePath = path.join(process.cwd(), 'lib', 'templates', 'course_template.json');
-      const courseTemplate = fs.readFileSync(templatePath, 'utf-8');
-      
-      systemPrompt = `You are an expert educational course designer. When the user asks to create a course, you immediately generate a complete, well-structured course based on their request.
 
-When the user requests to create a course:
-1. DO NOT ask questions or use the Socratic method
-2. DO NOT request additional information
-3. IMMEDIATELY create a complete course structure based on what they asked for
-
-Extract from their request:
-- The topic they want to learn
-- Implied skill level (if not specified, assume Beginner)
-- Any specific areas they mentioned
-
-Then create a course following this template:
-
-${courseTemplate}
-
-Fill in the template with:
-- Relevant, practical content based on the topic
-- Clear learning objectives
-- Mix of theory and practice lessons (focus on practice)
-- Progressive difficulty
-- Practical exercises and reflection questions
-- All content should follow the Socratic method for the actual learning (but NOT for course creation)
-
-Generate the course immediately and present it in this format:
-
-Я создал для вас курс "[название курса]". Вот его структура:
-
-[Brief description of what the course covers]
-
-Then output the complete course structure wrapped in:
-<COURSE_JSON>
-{your generated course JSON here}
-</COURSE_JSON>
-
-Remember: Generate the course IMMEDIATELY based on their request. Do not ask questions.`;
+      systemPrompt = COURSE_CREATION_PROMPT;
     }
     
     // Добавляем системный промпт
